adds in error handling to RepoCollection for assigning repos, starts testing for repo class
