+++ spec/services/github_adapter_spec.rb
@@ -198,21 +198,6 @@ RSpec.describe GithubAdapter do
     end
   end
 
-  describe "#convert_to_repos" do
-    before(:each) do
-      repos = adapter.client.repos(github_login)
-      @converted_repos = adapter.convert_to_repos(repos)
-    end
-
-    it "returns an array", :vcr do
-      expect(@converted_repos).to be_an_instance_of Array
-    end
-
-    it "converts an array of sawyers resource to an array of repo objects", :vcr do
-      expect(@converted_repos.first).to be_an_instance_of Repo
-    end
-  end
-
 
   describe "gist methods" do 
     let(:since) { "&since=#{two_weeks_ago}"}
